requirements.txt
setuptools~=68.2.0
boto3~=1.34.45
certbot~=2.9.0
cloudflare~=2.19.0
dnspython~=2.6.1